Added limited functionality for users
Added the ability to create an account and push it to the database.
Added the ability to login to the program using data from the db.
Added the ability to delete a user from the database.

Need to handle exceptions for login if the username/password
combo is incorrect.

Also need to give limited deletion access to non admin accounts
and keep that ability for the admin accounts.

Will need to refactor some code.

TODO:
1. Add account functionality
2. Implements accounts in both the program and the database
3. Connect accounts to their parent users.